Removing Dangerous and Unused Mod links
Some of these links on profiles are not confirmed and make looking at
profile pages on phones dangerous for mods and admins. Some like mod
report features are not used at all and just clutter the options. The
forum moderator links are no longer needed after the migraion to the new
forum, and are now misleading.