Changing Game Passwords to Invite Codes (#298)
Removing text referring to a game password and replacing it as an invite
code so people no longer think they need to enter their account password
to join a game. Also noticed the country progress color bar was missing
a css 2% margin making it stick out so fixed that too.