Various Anon Updates (#323)
* Various Anon Updates

After the additional anon protection various members complained it was
now difficult to easily see who had retreat or build options on large
maps. This change will replace the !!, !, grey check, and green check
options with the old forum lock icon to indicate who has moves both in
game and on the member home preview. Only game directors will be able to
see through the lock on the home preview (since TD's spectating games is
quite normal) whereas mods spectating games should not need to see
through (I'm planning a future enhancement to hide members in anon games
for mods who are spectating the game).

* Cover Left and None orders in the Anon Message

Extra check for Left and build/retreat phases. This is the final set of
possible circumstances for this message.