Add mod tool for country replacing (#342)
* Add Temp Ban Tool for Mods from vDip

Allows a moderator to specifiy a lockout time period for members in days
during which they cannot create or join games.

* Adding missed file

Forgot to check the last file.

* Add a new tool for fast player swaps

This tool will allow the mods to quickly swap people from 1 or all
games. When all games is selected, the 2 players will be swapped for all
the games where the replacement player met the RR requirement and where
the replacing player was not already in the game. A list will be shown
after to the moderator with the list of successful and failed swaps, and
why they failed if unsuccessful.