Emergency Fix for Issue from PHP Upgrade
I was able to narrow the problem down to the specific line:
$LoadedVariants[$variantID] = libVariant::loadFromVariantID($variantID);

The people on google who had the same problem discussed an issue around
serialization after PHP upgrades to version 7 which was just done.
Looking into the failing function libVariant::loadFromVariantID I see it
is also doing serialization, which is far out of my skill set in php. I
did confirm that the other places on the site that use this function are
working as intended. So for the time being I commented out this section
of the settings page and confirmed it no longer crashes for people who
have country mutes.