Fix 10 year Issue
The reset min bet function was never right for being called as a user was being removed from a pre-start game. The way the logic worked meant the user count wouldn't refect the real user count in the members table. Only discovered when the resetminbet function began getting called when it was initially supposed to. Count is now being pulled direct from the db to avoid state issues.