Add Emergency Pauses
Allow 1 automatic pause of all games for players who have completed at least 10 games and have not used an automatic pause in the last 6 months. Adds a mod tool to reset the time if a player presents a valid excuse or the option to ban members from using this tool at all. Mods can see on the profile page if a member qualifies for an automatic pause or not. This tool pauses the games and then emails the moderator team so they can follow up and post an explaination in the game chat.