Adjust Points for Bots
Adjust points so that bots will never have points taken or awarded. Keep the calls to the user function that controls point distribution because those functions control the member and game table records being updated for bet and pot. So a bot's "bet" in a game will always be 5 points and when they leave a game it will always remove 5 points from the game pot. This is because games with bots playing will have to be unranked, which will be forced in game creation options in another PR.