Minor additions to into, and fixed broken FAQ link
As per the discussion in Issue  #335, this commit makes some minor changes to the Intro to Diplomacy page, adding some clarifications to common misconceptions new players have about the support hold order. The changes are minor, since there was resistance to making the page overly complex.
Also adds a link to the official Avalon Hill rulebook on the intro page, and corrects a broken link to the rulebook on the FAQ.