Publish a mod only Alpha View New Profile
This is an alpha for the page that will eventually replace the exisiting profile page. This page will be designed for sites that have gotten rid of the built in forum, the old profile page will be left intact for vDip or other implimentations that do not have a custom forum.