Force maps to fit their container
This fix forces the map to fit its container, preventing it from overhanging to the left.
It also autoshrinks oversized maps to fit. Maps that are excessively large may not remain legible on the inpage map. (On the variants webDip currently has, this is a non-issue. But smallMaps over ~1100px wide will become increasingly difficult to read.)